Add metropolis_hamiltonian kernel in numba
