+++ netket/sampler/__init__.py
@@ -1,5 +1,6 @@
 from .abstract_sampler import AbstractSampler
 from .metropolis_hastings import *
-from .metropolis_local import *
-from .metropolis_exchange import *
+from .metropolis_local import MetropolisLocal, MetropolisLocalPt
+from .metropolis_exchange import MetropolisExchange, MetropolisExchangePt
+from .metropolis_hamiltonian import MetropolisHamiltonian, MetropolisHamiltonianPt
 from .exact_sampler import *
